separate app/gfx refactor
