+++ src/sokol/c/sokol_debugtext.c
@@ -1,3 +1,3 @@
+#include "sokol_defines.h"
 #include "sokol_gfx.h"
 #include "sokol_debugtext.h"
-
