+++ src/sokol/c/sokol_gfx.c
@@ -1,2 +1,2 @@
 #include "sokol_defines.h"
-#include "sokol_audio.h"
+#include "sokol_gfx.h"
