+++ src/sokol/c/sokol_gl.c
@@ -1,2 +1,3 @@
+#include "sokol_defines.h"
 #include "sokol_gfx.h"
 #include "sokol_gl.h"
