+++ src/sokol/c/sokol_time.c
@@ -1,1 +1,2 @@
+#include "sokol_defines.h"
 #include "sokol_time.h"
